refactor(phase-10F): move all routers into api/routers/
Twelve router modules move out of openrag/routers/ into
openrag/api/routers/{user,admin,auth}/, and a new
api/routers/user/health.py takes over the /health_check + /version
endpoints that lived inline in api/main.py.

Layout
------
api/routers/user/
  chat.py         <- routers/openai.py
  extract.py      <- routers/extract.py
  health.py       (new — /health_check, /version)
  search.py       <- routers/search.py

api/routers/admin/
  cluster.py      <- routers/actors.py
  indexing.py     <- routers/indexer.py
  jobs.py         <- routers/queue.py
  monitoring.py   <- routers/monitoring.py (already partly moved in 10C)
  partitions.py   <- routers/partition.py
  tools.py        <- routers/tools.py
  users.py        <- routers/users.py
  workspaces.py   <- routers/workspaces.py

api/routers/auth/
  oidc.py         <- routers/auth.py

The auth split planned by the strategy doc (login.py + oidc.py)
collapses to a single oidc.py for now: OpenRAG has no token-login
endpoint today, every route in the source file is OIDC-flow or the
protected /auth/me debug route. login.py can be added later if a
token-mode login endpoint ever lands.

Strangler-fig shims
-------------------
Each legacy openrag/routers/<file>.py is rewritten to:

    from api.routers.<area>.<file> import *
    from api.routers.<area>.<file> import router
    __all__ = ["router"]

so chainlit_api.py, the still-active openrag/main.py entrypoint
(uvicorn main:app), and the existing routers/test_*.py tests keep
working through Phase 12 cleanup. routers/monitoring.py also keeps
the MonitoringMiddleware = InstrumentationMiddleware alias alive for
the legacy main.py boot path.

api/main.py
-----------
Switches every ``from routers.<file> import router`` to
``from api.routers.<area>.<file> import router``, drops the inline
/health_check and /version routes (now mounted via health_router),
and keeps the same prefixes / tag set / WITH_OPENAI_API /
WITH_CHAINLIT_UI gating so the OpenAPI schema is byte-identical to
before the move.

Test patch fix
--------------
openrag/test_token_validation.py patches ``routers.openai.get_num_tokens``;
the symbol now lives in ``api.routers.user.chat`` (patching a name in
the shim's namespace does not intercept calls made from the new module,
so the patch target follows the definition). Two assertions updated;
all 10 tests in the file pass.

Verified: 944 tests pass across openrag/ (excluding the heavy
components/indexer integration subtree); ruff clean.
