+++ openrag/routers/monitoring.py
@@ -1,32 +1,24 @@
-"""Prometheus ``/metrics`` endpoint for OpenRAG.
+"""Legacy import path for the monitoring router (Phase 10F shim).
 
-The middleware that powered this used to live alongside the endpoint
-here; Phase 10C moved it to :mod:`api.middleware.instrumentation`
-because middleware is request infrastructure, not a router concern. The
-``/metrics`` endpoint itself still lives in this module — 10F moves it
-to ``api/routers/admin/monitoring.py``. Until then ``MonitoringMiddleware``
-is re-exported below so the legacy ``openrag/main.py`` import keeps
-working through the strangler-fig window.
+The ``/metrics`` endpoint now lives at
+:mod:`openrag.api.routers.admin.monitoring`. Phase 10C had already
+moved the middleware to :mod:`openrag.api.middleware.instrumentation`
+while leaving the endpoint here; 10F finishes the move and turns this
+file into a re-export shim.
+
+``MonitoringMiddleware`` is re-exported as an alias for the
+:class:`InstrumentationMiddleware` rename so the legacy
+``openrag/main.py`` keeps booting through the strangler-fig window;
+Phase 12 cleanup deletes both this shim and the alias.
 """
 
-import asyncio
-
 from api.middleware.instrumentation import InstrumentationMiddleware
-from fastapi import APIRouter, Depends
-from fastapi.responses import Response
-from routers.utils import require_admin
-from utils.monitoring import get_metrics
+from api.routers.admin.monitoring import *  # noqa: F401,F403
+from api.routers.admin.monitoring import router
 
-router = APIRouter()
-
-
-# Legacy alias for ``openrag/main.py``; Phase 12 cleanup removes it once
-# the last importer flips to ``api.middleware.instrumentation``.
+# Legacy name kept alive for the still-active ``openrag/main.py``
+# entrypoint; the new ``openrag/api/main.py`` imports the renamed
+# class directly from ``api.middleware``.
 MonitoringMiddleware = InstrumentationMiddleware
 
-
-@router.get("/metrics", summary="Prometheus metrics endpoint", dependencies=[Depends(require_admin)])
-async def prometheus_metrics():
-    """Return all metrics in Prometheus text exposition format."""
-    content = await asyncio.to_thread(get_metrics)
-    return Response(content=content, media_type="text/plain; version=0.0.4; charset=utf-8")
+__all__ = ["router", "MonitoringMiddleware"]
